requirements.txt
amqp==5.0.9
asgiref==3.5.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.12
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-filter==21.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
importlib-metadata==4.11.0
kombu==5.2.3
Markdown==3.3.6
packaging==21.3
pkg_resources==0.0.0
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
redis==4.1.3
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zipp==3.7.0
